Remove timestamp synchronization and RTT computation
Remove all logic for collecting and synchronizing the 4 timestamps (TX
and RX on client and server side). Instead, simply log each packet
timestamp as soon as it is available locally, leaving matching of all
the timestamps to post processing. This greatly simplifies the logic
and removes several shared queues and their associated
locks. Furthermore, this removes the need for the server to transmit
its timestamps to the client.

To still allow post processing of the server timestamps, support the
--log option on the server side as well. Rework the logging mechanism
to be part of the nanoping_instance (and nanoping_init()) instead of
only handled by the client side code. Furthermore, change the logging
to append to files (as would be expected for logs) instead of
overwriting them.

Furthermore, remove the __packed__ attribute from the
nanoping_emul_txs struct. This struct is only used locally, so the
exact format of it should not be of critical importance, i.e. any
unexpected padding from the compiler to align the struct members
should have no bearing on its function. Furthermore, the members of
the struct are already aligned, so packing the structure does
nothing. Therefore, remove the __packed__ attribute to ensure that the
struct is aligned to avoid any potential issues with accessing
unaligned structs.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>