Remove unnecessary fields from packet
Since simplifying nanoping to no longer send the remote timestamps in
the packets, remove these now unused fields to allow for a smaller
packet size.

Currently only 9 bytes are required for the packet message format (8
bytes for the sequence numbers, 1 byte for type). However, instead of
using the __packed__ attribute to create an unaligned 9-byte
struct (which may harm performance and cause issues), manually pad it
so it becomes 8-byte aligned by reserving 7 bytes for future use. As
the data is sent over the network it is important it has the expected
format, so insert a compile time check to help detect if any future
changes causes the compiler to insert unexpected padding.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>