Add an option to use a busy-loop instead of sleep to wait
Add an alternative way to wait for the next sending interval, using a
busy-loop instead of sleep. The busy-loop provides a more accurate
wait, espeically for delays < 100us, where the sleep is regularly
overshoots the intended wait period, at the cost of high CPU
utilization.

Add the --timer/-T option to configure which "timer" mechanism to use
for sending packets at the specified interval. The "sleep" option
corresponds to the previous way of doing it (usleep()), while the
"busy" option uses the new busy-loop mechanism.

Performing a test with --delay=10 (us), the sleep timer achieves an
average IPD of 69.9 us (MAE 60 us), while the busy timer achieves an
average IPD of 10.3 us (MAE 0.9 us).

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>