Make SYN-packet send test configuration
Make the necessary backend changes for supporting sending and
receiving custom payload with the nanoping_send_one() and
nanoping_receive_one() functions.

Use the added capability to send custom payloads to send a test
configuration in the initial SYN packet which the server can
access. These test configuration options can later be used by the
server when the reverse mode is added.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>