Add reverse mode
Add the -R/--reverse option, which makes the server send pings to the
client instead (and the client echo the pings). Make the server use
the client's settings for delay, timer etc. that are sent in the
initial SYN packet. Have the server santize the received client
settings, overriding them if they are outside of expected ranges.

Note that the server will switch into "client mode" (sender) as soon
as it has ACKed the initial SYN, and will therefore not be able to
send additional SYN-ACKs if the client does not receive the initial
SYN-ACK. A loss of the initial SYN-ACK will therefore cause the
reverse test to fail. While not ideal, this used to be an issue for
nanoping in regular mode as well until it was recently updated to
check the client address, so I deem this flaw acceptable for now.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>