Minor refactor of client sendloop arguments
Instead of sending the delay, timer-type and if a handshake should be
performed or not as individual arguments to the client_sendloop()
function, simply send the nanoping_client_opts instance that already
groups all of these together. This reduces the clutter in the
arguments and makes the code a bit less painful to read.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>