Change header offset calculation in tx timestamp processing
For the tx timestamp processing, the timestamp together with the full
packet the timestamp is for is received on the error queue. This full
packet includes the Ethernet, IP, and UDP headers, which the
application is generally not aware of (and just so happen to sum up to
the magical number 42).

To find the nanoping message header (which is required to associate
the timestamp with the correct nanoping sequence number), the offset
into the packet buffer to look for the nanoping header was previously
calculated on the basis of the expected size of the nanoping header
and payload. This approach, however, will not work if nanoping is to
support sending packets of different sizes (and already does not work
for the initial SYN-packet, which may be of a different
size). Therefore, change the calculation to instead assume the size of
the underlying network headers (Ethernet + IP + UDP) and use that as
the basis to find the offset into the packet buffer where the nanoping
header is found.

Preferably the sizes of the underlying network headers should not be
assumed, but I was unable to find any mechanism to query kernel for
this information. Nanoping currently only works over UDP and IPv4, and
therefore this assumption should be relatively safe. In theory I guess
nanoping may work over other link layers than Ethernet, but I'm not
sure any non-Ethernet interfaces support hardware timestamps, and is
not something that is of interest for us to support at the moment.

An alternative implementation would be to actually parse the entire
header-stack in the packet (i.e. parse the initial Ethernet-header,
check that next header is IPv4, parse IPv4 header and check it's IHL
size etc.). While this would perhaps be slightly more robust (although
also introduce some additional error sources), it adds additional
overhead and complexity for very limited benefits compared to simply
assuming the header size.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>