Add pong-every N option
Add the -y/--pong-every option to control the rate at which ping
packets will be ponged. The default of 1 is to pong every ping, but if
set to N then only every Nth ping (or specifically every Nth sequence
number) will be ponged.

Furthermore, pong-every can be set to 0 to entierly disable ponging so
that only pings are sent, and outside of the handshake packets will
then only be sent one way.

Note that by default both the client and the server time out if they
do not receive any packets for 5 seconds. To avoid the lack of pongs
from causing a timeout you can increase the timeout value using the
-t/--timeout option (has to be done for both the server and the
client).

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>