Rename nanoping to nanoprobe
At this point we have made extensive changes to our nanoping
fork. Among other, the tool can no longer be used the same way, as it
no longer tries to match up packets to compute the RTTs itself (only
dumps it to a log for post processing). To distinguish our heavily
reworked fork of nanoping from the original nanoping, rename it to
nanoprobe.

Only do a "surface level" remaining, i.e. rename the source files and
the compiled binary, but not all struct and function names within the
source code. Thereby, keep the heritage from nanoping clear in for
those reading the source code. In the source code, only rename the
`nanoping_client_opts` struct to `nanoprobe_client_opts`, as that is
the sole struct with "nanoping" in its name that has not been
inherited from the original nanoping. Also change all instances of
"nanoping" in the produced output to "nanoprobe".

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>