Verify sender in client receive thread
For messages received by the client receive thread, verify that they
are from the expected sender (i.e. the server).

Previously, the client receive thread did not check who the received
packets were from. This could result in it acting on packets from
another nanoprobe client/server instance than the one it is connected
to.

While fixing this is good in its own right, it becomes especially
important in the context of a future commit that will also make the
client receive thread react to ACK packets. If the source of the
packets is not the expected sender then, it could allow other nanoping
clients to successfully connect to a an already connected nanoping
client.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>