Resend SYN-ACK in reverse/duplex modes
When running in reverse or duplex mode, the server would switch to
"client-mode" directly after completing the handshake and sending a
SYN-ACK to the client. If this SYN-ACK was lost, the client would
never be able to complete the handshake as the server would no longer
respond to any resent ACKs once in "client-mode".

To fix this, make it possible for the server to resend SYN-ACKs (upon
receiving additional ACKs from the same client) even when in
"client-mode". Therefore, add handling of SYN messages to the client
receive loop (if the nanoprobe instance was originally in server
mode).

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>