Add output to indicate if connection succeeds/fails
While the server already prints out a message once it successfully
connects to the client, the client does not report when or if the its
handshake with the server succeeds or fails. Therefore, add some
additional output to make it clear whenever the connection
establishment succeeds or fails on both the client and server side.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>