Remove initial wait in client receive thread
The initial wait for any received packet in the client does not make a
lot of sense to me. It is not really needed, as the client will
initiate the connection towards the server immediately upon startup,
so risk for unintended timeout between creation of the receive thread
and the starting the connection should be minimal. Furthermore, this
wait prevents the initial client handshake from ever timing out if it
does not get any response from the server (as then it will wait
indefinitely for the first packet to arrive), preventing the normal
timeout behavior from killing it. Therefore, remove this initial wait.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>