Remove dummy-pkt option
The dummy-pkt option is a remnant from nanoping that has not really
been maintained throughout nanoprobes evolution. While it still
somewhat works, it does not necessarily behave as expected (being set
separately for client/server, does not respect the ping/pong
size). Furthermore, due to using a completely different socket, that
does not take part in the handshake, it will not be able to pass
through NAT (for reverse/duplex modes) and triggers warnings on the
receiving side due to being from a non-connected port.

As asymmetrical bandwidth can now be achieved by either using
different ping/pong sizes, or with the pong-every option, it seems
unlikely that we will have much use for this option.

So to shrink the code base and reduce future maintenance and
complexity issues as nanoprobe evolves, remove this feature.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>