Drop address from error queue receive
When receiving the TX timestamps from the error queue, we also
requested the source address of the packet that generated the
error, but never used it after.

As we do not use the src address in this context (and should already
know that it will be from ourself), stop requesting this unecessary
information. This might slightly reduce overhead.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>