Add option to busy-loop while waiting for HW TX timestamps
As reading the socket error-queue is a non-blocking operation, the
code for retrieving a TX timestamp (in nanoping_txs_one()) uses
select() to be notified when there's an error message to be
read. While this works well at lower packet rates, it may fail to keep
up at higher packet rates.

Therefore, add the option -B/--busyloop-txtstamp, which will instead
read the error queue in a busy loop until there is a message
available. If the option is not passed, select() will be used as
previously.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>