Small code style fix
In the argument-parsing for the --probe-schedule option there was a
rather gnarly-looking multi-line if-condition that I assume I wrote
because I didn't want to allocate a variable for holding the return
value. As there's now a suitable err variable already available,
rewrite this into a separate function call with the if-condition only
checking the value of the err variable, making the actual condition
much clearer.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>