Add option to pin each thread to its own CPU core
Add the -P/--pin-threads <x,y,z> option that let's the user specify
which 3 cores to pin each of the nanoprobe threads to. Pinning the
threads to separate cores can help slightly in high packet rates (low
delays) scenarios as it avoids any interfering effect from CPU
migration.

The x-core handles the main thread, running the client or server loop,
the y-core handles the thread fetching TX timestamps, and the z-core
runs the client receive thread (receiving and logging pong
messages). The client-receive thread will not run at all for the
server in the forward mode, for the client in the reverse mode, and
for neither in the duplex mode. Technically nanoprobe may also have a
4th thread for handling the signal to stop after the count timeout,
but this is such a low activity thread that I see no reason to pin it.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>